fix bug in step size

+++ surfaces/mathematical_functions/test_functions_1d/gramacy_and_lee_function.py
@@ -39,7 +39,7 @@ class GramacyAndLeeFunction(MathematicalFunction):
     def search_space(self, value_types="array", steps=100):
         min_x0 = 0.5
         max_x0 = 2.5
-        step_size_x0 = int((max_x0 - min_x0) / steps)
+        step_size_x0 = (max_x0 - min_x0) / steps
 
         return super().search_space(
             search_space_blank={
