+++ surfaces/mathematical_functions/test_functions_2d/beale_function.py
@@ -55,8 +55,8 @@ http://www-optima.amp.i.kyoto-u.ac.jp/member/student/hedar/Hedar_files/TestGO
         max_x0 = 4.5
         max_x1 = 4.5
 
-        step_size_x0 = int((max_x0 - min_x0) / steps)
-        step_size_x1 = int((max_x1 - min_x1) / steps)
+        step_size_x0 = (max_x0 - min_x0) / steps
+        step_size_x1 = (max_x1 - min_x1) / steps
 
         return super().search_space(
             search_space_blank={
