+++ surfaces/mathematical_functions/test_functions_2d/eggholder_function.py
@@ -46,8 +46,8 @@ class EggholderFunction(MathematicalFunction):
         max_x0 = 1000
         max_x1 = 1000
 
-        step_size_x0 = int((max_x0 - min_x0) / steps)
-        step_size_x1 = int((max_x1 - min_x1) / steps)
+        step_size_x0 = (max_x0 - min_x0) / steps
+        step_size_x1 = (max_x1 - min_x1) / steps
 
         return super().search_space(
             search_space_blank={
